% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvec.R
\name{bvec}
\alias{bvec}
\title{Function to compute m by 1 column vector \eqn{I} (independent and dependent
cases) of integrated emulator with squared exponential kernel.}
\usage{
bvec(
  w,
  z = NULL,
  mu,
  Sigma,
  z_design = NULL,
  params,
  ind = "TRUE",
  det_comp = NULL,
  LS.inv = NULL
)
}
\arguments{
\item{w}{an \eqn{m} (number of design points) by \eqn{d} (dimensions)
matrix of design points.}

\item{z}{a vector of length d1 of exogenous inputs.}

\item{mu}{a vector of predictive mean values of length k.}

\item{Sigma}{a variance-covariance matrix.}

\item{z_design}{a design matrix of exogenous inputs.}

\item{params}{a list with parameter values.}

\item{ind}{logical argument indicating if we adopt independence assumption.}

\item{det_comp}{a determinant component in the calculations.}

\item{LS.inv}{an inverse of the sum of covariance matrix and Lambda.}
}
\value{
a vector of covariances between input of interest and design point.
}
\description{
Function to compute m by 1 column vector \eqn{I} (independent and dependent
cases) of integrated emulator with squared exponential kernel.
}
