% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covar_pred}
\alias{covar_pred}
\title{Predictive variance-covariance for w of the integrated emulator.}
\format{
A tibble with 36 rows and 6 variables:
\describe{
\item{var1}{predictive variance for quarterly heating demand.}
\item{var2}{predictive variance for a quarterly gas price.}
\item{var3}{predictive variance for a quarterly electricity price.}
\item{covar12}{predictive covariance between x1 and x2 inputs.}
\item{covar13}{predictive covariance between x1 and x3 inputs.}
\item{covar23}{predictive covariance between x2 and x3 inputs.}
}
}
\source{
outputs of GP emulation and two DLMs.
}
\usage{
covar_pred
}
\description{
A dataset containing the values of variance and covariances for w.
}
\keyword{datasets}
