% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovMatrix.R
\name{CovMatrix}
\alias{CovMatrix}
\title{Constructs a squared exponential correlation matrix}
\usage{
CovMatrix(Design, cls)
}
\arguments{
\item{Design}{a matrix of inputs for design set}

\item{cls}{a vector of correlation length parameters}
}
\value{
A matrix of correlation calculated for design points
}
\description{
Constructs a squared exponential correlation matrix
}
