% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jmatrix.R
\name{Jmatrix}
\alias{Jmatrix}
\title{Function to compute \eqn{J} matrix of Linked emulator with
squared exponential correlation function.}
\usage{
Jmatrix(w, z = NULL, mu, Sigma, z_design = NULL, params, ind = "TRUE")
}
\arguments{
\item{w}{an \eqn{m} (number of design points) by \eqn{d} (dimensions)
matrix of design points}

\item{z}{a vector of length d1 of exogenous inputs.}

\item{mu}{a vector of length \eqn{d} of predictive mean.}

\item{Sigma}{a \eqn{d} by \eqn{d} variance-covariance matrix.}

\item{z_design}{a design matrix of exogenous inputs.}

\item{params}{a list with parameter values.}

\item{ind}{logical argument indicating if we adopt independence
assumption for \eqn{W}.}
}
\value{
an \eqn{m} by \eqn{m} matrix \eqn{J}.
}
\description{
Function to compute \eqn{J} matrix of Linked emulator with
squared exponential correlation function.
}
