% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_fun.R
\name{mean_fun}
\alias{mean_fun}
\title{Function to compute the analytical mean \eqn{mu_L} of
the Linked Emulator under squared exponential correlation function}
\usage{
mean_fun(
  w,
  z = NULL,
  y,
  mu,
  Sigma,
  z_design = NULL,
  params,
  FastParts,
  ind = "TRUE"
)
}
\arguments{
\item{w}{an \eqn{m} (number of design points) by \eqn{d} (dimensions)
matrix of design points}

\item{z}{a vector of length d1 of exogenous inputs.}

\item{y}{a vector length \eqn{m} (number of design points) responses.}

\item{mu}{a vector of length \eqn{d} of predictive mean.}

\item{Sigma}{a \eqn{d} by \eqn{d} variance-covariance matrix.}

\item{z_design}{a design matrix of exogenous inputs.}

\item{params}{a list with parameter values.}

\item{FastParts}{a list of precomputed objects.}

\item{ind}{logical argument indicating if we adopt independence
assumption for \eqn{W}.}
}
\value{
a value of the analytical mean.
}
\description{
Function to compute the analytical mean \eqn{mu_L} of
the Linked Emulator under squared exponential correlation function
}
