% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_pred}
\alias{mean_pred}
\title{Predictive means for w of the integrated emulator.}
\format{
A tibble with 36 rows and 3 variables:
\describe{
\item{mu1}{predictive mean for quarterly heating demand}
\item{mu2}{predictive mean for a quarterly gas price.}
\item{mu3}{predictive mean for a quarterly electricity price.}
}
}
\source{
outputs of GP emulation and two DLMs.
}
\usage{
mean_pred
}
\description{
A dataset containing the values of predictive means for w.
}
\keyword{datasets}
