% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_fun.R
\name{psi_fun}
\alias{psi_fun}
\title{Fuction to compute \eqn{\psi_{jk}} for d inputs of integrated emulator
for (independent case) and (dependent case) with a squared exponential kernel.}
\usage{
psi_fun(w_j, mu, Sigma, params, ind = "TRUE", Lambda.mat = NULL, LS.inv = NULL)
}
\arguments{
\item{w_j}{a vector of size k.}

\item{mu}{a vector of predictive mean values of length k.}

\item{Sigma}{a variance-covariance matrix.}

\item{params}{a list with parameter values.}

\item{ind}{logical indicating if we adopt independence assumption.}

\item{Lambda.mat}{Lambda matrix.}

\item{LS.inv}{logical indicating if we adopt independence assumption.}
}
\value{
a vector.
}
\description{
Fuction to compute \eqn{\psi_{jk}} for d inputs of integrated emulator
for (independent case) and (dependent case) with a squared exponential kernel.
}
