% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{w_valid}
\alias{w_valid}
\title{Validation (test) matrix for w of the integrated emulator.}
\format{
A tibble with 36 rows and 3 variables:
\describe{
\item{x1}{input of w matrix, quarterly heating demand}
\item{x2}{input of w matrix, a quarterly gas price.}
\item{x3}{input of w matrix, a quarterly electricity price.}
}
}
\source{
latin hypercube
}
\usage{
w_valid
}
\description{
A dataset containing the values of validation matrix w.
}
\keyword{datasets}
