% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_fun.R
\name{xi_fun}
\alias{xi_fun}
\title{Function to compute \eqn{\xi_{ik}} (independent case) and \eqn{\xi_{i}}
(dependent case) of integrated emulator with a squared exponential kernel.}
\usage{
xi_fun(w_i, mu, Sigma, params, ind = "TRUE", det_comp = NULL, LS.inv = NULL)
}
\arguments{
\item{w_i}{a vector of size k.}

\item{mu}{a vector of predictive mean values of length k.}

\item{Sigma}{a variance-covariance matrix.}

\item{params}{a list with parameter values.}

\item{ind}{logical indicating if we adopt independence assumption.}

\item{det_comp}{a determinant component in the calculations.}

\item{LS.inv}{an inverse of the sum of covariance matrix and Lambda.}
}
\value{
a vector (independent case) or a single value (dependent case).
}
\description{
Function to compute \eqn{\xi_{ik}} (independent case) and \eqn{\xi_{i}}
(dependent case) of integrated emulator with a squared exponential kernel.
}
