% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y_design}
\alias{y_design}
\title{Output vector of the energy system model.}
\format{
A vector of size 120:
\describe{
\item{y}{a vector containing the simulations.}
}
}
\source{
output of simulations
}
\usage{
y_design
}
\description{
A vector containing the simulations of energy system model.
}
\keyword{datasets}
