% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y_valid}
\alias{y_valid}
\title{Output vector of the energy system model used for validation.}
\format{
A vector of size 36:
\describe{
\item{y}{a vector containing the simulations.}
}
}
\source{
output of simulations
}
\usage{
y_valid
}
\description{
A vector containing the simulations of energy system model.
}
\keyword{datasets}
