% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z_design}
\alias{z_design}
\title{Design matrix for z of the integrated emulator.}
\format{
A tibble with 120 rows and 2 variables:
\describe{
\item{z1}{input of z matrix, efficiency of gas boiler.}
\item{z2}{input of z matrix, efficiency of heat pump.}
}
}
\source{
latin hypercube
}
\usage{
z_design
}
\description{
A dataset containing the values of design matrix z.
}
\keyword{datasets}
