% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{z_valid}
\alias{z_valid}
\title{Validation (test) matrix for z of the integrated emulator.}
\format{
A tibble with 36 rows and 2 variables:
\describe{
\item{z1}{input of z matrix, efficiency of gas boiler.}
\item{z2}{input of z matrix, efficiency of heat pump.}
}
}
\source{
latin hypercube
}
\usage{
z_valid
}
\description{
A dataset containing the values of validation matrix z.
}
\keyword{datasets}
