% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta_fun.R
\name{zeta_fun}
\alias{zeta_fun}
\title{Function to compute \eqn{\zeta_{ijk}} (independent case) and \eqn{\zeta_{ij}}
(dependent case) of integrated emulator with a squared exponential kernel.}
\usage{
zeta_fun(
  w_i,
  w_j,
  mu,
  Sigma,
  params,
  ind = "TRUE",
  det_comp = NULL,
  Gamma.inv = NULL,
  GS.inv = NULL
)
}
\arguments{
\item{w_i}{a vector of size k.}

\item{w_j}{a vector of size k.}

\item{mu}{a vector of predictive mean values of lenght k.}

\item{Sigma}{a variance-covariance matrix.}

\item{params}{a list with parameter values.}

\item{ind}{logical indicating if we adopt independece assumption.}

\item{det_comp}{a determinant component in the calculations.}

\item{Gamma.inv}{an inverse of Gamma matrix.}

\item{GS.inv}{an inverse of the sum of covariance matrix and Gamma.}
}
\value{
a vector (independent case) or a single value.
}
\description{
Function to compute \eqn{\zeta_{ijk}} (independent case) and \eqn{\zeta_{ij}}
(dependent case) of integrated emulator with a squared exponential kernel.
}
